+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0] - 2022-08-04
 ### Added
 - `Records` now have `fileName` which has the full filename associated with a record.
 - The `getRecord` function, which will return a record associated with an archive.
@@ -26,7 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `Archive` type.
 - The `ClientConfig` type which allows an sdk user to specify their credentials and other configuration settings.
 
-[Unreleased]: https://github.com/permanentorg/permanent-sdk/base/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/permanentorg/permanent-sdk/base/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.3.0
 [0.2.1]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.2.1
 [0.2.0]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.2.0
 [0.1.0]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.1.0